+++ com.unity.mobile.android-logcat/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Provide user friendly message when NDK directory is incorrect or is not set.
  - Don't show Add/Remove Tag menu item, if tag is empty.
  - Don't show Filter By Process Id menu item, if process id is invalid.
+ - When using AutoRun, logcat will appear automatically if you're building only for Android, previously it would appear for any platform, which is undesired.
 
 ## [1.2.1] - 2021-01-21